kernel: sched() should not imply spllo()
The idea is:

When we call sched() with interrupts disabled,
it must not return with them re-enabled.

The code becomes easier to reason about if
we make sched() preserve interrupt status,
which lets us call sched() from an interrupt
handler (with interrupts disabled) without
risking preemption by another interrupt once
sched() returns which can pump-up the stack.

This allows removing Proc.preempted flag as
it is now impossible for interrupts to
preempt each other in preempted().

Extra cleanups:

make interrupted() _Noreturn void
and remove unused Proc.yield flag.

+++ sys/src/9/port/taslock.c
@@ -106,7 +106,7 @@ lock(Lock *l)
 void
 ilock(Lock *l)
 {
-	ulong x;
+	int x;
 	uintptr pc;
 
 	pc = getcallerpc(&l);
